pyproject.toml
[project]
name = "voice-assistant-web"
version = "0.1.0"
description = "Real-time voice assistant demo with Soniox STT and ElevenLabs TTS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Flask==3.0.0",
  "flask-sock==0.7.0",
  "simple-websocket==1.0.0",
  "websocket-client==1.7.0",
  "httpx==0.27.0",
  "python-dotenv==1.0.1",
  "openai==1.40.0",
  "langdetect==1.0.9",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.6",
  "ruff>=0.5.0",
]


